pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "materials-project-ml"
version = "0.1.0"
description = "Matbench crystal graph benchmark with CGCNN, ALIGNN, and M3GNet"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alignn>=2025.4.1",
    "dgl>=2.1,<2.2",
    "jarvis-tools>=2024.10",
    "matgl>=2.0.9",
    "matminer>=0.10",
    "matplotlib>=3.9",
    "numpy>=1.26",
    "pandas>=2.2",
    "pymatgen>=2025.1.9",
    "scikit-learn>=1.5",
    "seaborn>=0.13",
    "torch>=2.2",
    "torch-geometric>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "ruff>=0.9",
]

[project.scripts]
mp-crystal-ml = "mp_crystal_ml.pipeline:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
